requirements.txt
django==4.0.5
mysqlclient==2.1.0
djangorestframework==3.13.1
django-crispy-forms==1.14.0
django-tables2==2.4.1
django-filter==22.1